Always use Ruby's `Process.spawn` implementation
The `posix-spawn` gem was created to address performance issues in Ruby
when spawning child processes. Those issues have been addressed since
Ruby 2.2 so the gem is technically not necessary anymore.

The performance of Ruby's `Process.spawn` is equivalent and sometimes a
little faster than the custom `POSIX::Spawn#_pspawn` method created by
the `posix-spawn` gem.

Additionally, the `posix-spawn` C extension is not compatible with Ruby
3.0. In Ruby 3.0, `IO.pipe` and others return nonblocking pipes - see
https://github.com/ruby/ruby/commit/6a65f2b1e479f268489b51a004b6c153c634c68a.

As you can see in the commit above, these are moved back to blocking (at
least for `stdin`, `stdout` and `stderr`) inside Ruby's `Process.spawn`
implementation. When using `posix-spawn` this results in the child
process getting an `EAGAIN` sometimes, which it doesn't expect as it
doesn't expect pipes to be non-blocking. This seems to happen when
writing more than 64KB on Linux.

More details in https://github.com/rtomayko/posix-spawn/issues/90