Add Progeny::Command::spawn_with_pipes
This is a copy of POSIX::Spawn::popen4. Originally, I was hoping to
remove this implementation altogether in favor of using Open3's
`popen3()`. However, Open3 uses `Process.detach`, which is more than
`popen4` does and it does not set `$?` like `Process.waitpid` does,
so it made migration for `posix-spawn` users more complicated.